moved METag into custom_tags.py; updated references
